+++ src/cgeo/geocaching/cgTrackable.java
@@ -23,7 +23,7 @@ public class cgTrackable {
 	public Spannable nameSp = null;
 	public String type = null;
 	public Date released = null;
-	public Double distance = null;
+	public Float distance = null;
 	public String origin = null;
 	public String owner = null;
 	public String ownerGuid = null;
